requirements.txt
#azure==4.0.0
azure-storage==0.36.0
azure-storage-blob==12.2.0
azure-identity==1.3.0
pyyaml
filetype
